refactor: page folder separation
