+++ lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'widgets/appbar.dart';
-import 'widgets/body.dart';
-import 'widgets/floating_buttons.dart';
+import './pages/home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -17,52 +15,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         useMaterial3: true,
       ),
-      home: const MyHomePage(title: 'Contador'),
+      home: const Home(title: 'Contador'),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key, required this.title});
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
-  void _decrementCounter() {
-    setState(() {
-      if (_counter > 0) {
-        _counter--;
-      }
-    });
-  }
-
-  void _resetCounter() {
-    setState(() {
-      _counter = 0;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: bar(title: widget.title),
-        body: Body(counter: _counter),
-        floatingActionButton: FloatingButtons(
-          decrementCounter: _decrementCounter,
-          resetCounter: _resetCounter,
-          incrementCounter: _incrementCounter,
-        ));
   }
 }
